Calculate clone fate bias for specific cell type

Calculate clone fate bias for specific cell type

Usage

clone_fate_bias(data, fate_use = "")

Arguments

data

data.frame, indicating lineage tracing data, first column should be lineage tracing barcodes, second column should be related cell type

fate_use

character, targeted fate cell type. if fate_use == '', this function will automatically select clone dominant cell type.

Examples

clone_fate_bias(fate_test)
#> [1] "stromal cell"
#>            clone_name     fate_use clone_size         fate_ratio
#> I                   I stromal cell         12 0.0833333333333333
#> X1                  1 stromal cell         32            0.59375
#> X9                  9 stromal cell         66  0.484848484848485
#> X123              123 stromal cell         27  0.592592592592593
#> X1234789      1234789 stromal cell          2                  1
#> X56IHG          56IHG stromal cell          2                  1
#> C                   C stromal cell          5                0.2
#> X56789          56789 stromal cell          5                0.2
#> X3FE              3FE stromal cell          1                  1
#> IFE               IFE stromal cell          1                  1
#> X12G              12G stromal cell          1                  1
#> X145              145 stromal cell          1                  1
#> X1256789      1256789 stromal cell          1                  1
#> X12347          12347 stromal cell          1                  1
#> X123456789  123456789 stromal cell          1                  1
#> X12C69          12C69 stromal cell          1                  1
#> X12GFEDC89  12GFEDC89 stromal cell          1                  1
#> CB14GFE       CB14GFE stromal cell          1                  1
#> EDA               EDA stromal cell          1                  1
#> X7H9              7H9 stromal cell          1                  1
#> CB56789       CB56789 stromal cell          1                  1
#> CBAHG           CBAHG stromal cell          1                  1
#> X1HEDCB9      1HEDCB9 stromal cell          1                  1
#> X12345F9      12345F9 stromal cell          1                  1
#> C456IHG       C456IHG stromal cell          1                  1
#> X16G              16G stromal cell          1                  1
#> X125              125 stromal cell          1                  1
#> X123FED9      123FED9 stromal cell          1                  1
#> X12589          12589 stromal cell          1                  1
#> X3456789      3456789 stromal cell          1                  1
#> X589              589 stromal cell          1                  1
#> X12789          12789 stromal cell          1                  1
#> X14589          14589 stromal cell          1                  1
#> X12E              12E stromal cell          1                  1
#> AD5               AD5 stromal cell          1                  1
#> X567H9          567H9 stromal cell          1                  1
#> X5HGBA          5HGBA stromal cell          1                  1
#> X1234589      1234589 stromal cell          1                  1
#> X16I8G          16I8G stromal cell          1                  1
#> X14IHGFE      14IHGFE stromal cell          1                  1
#> A2349           A2349 stromal cell          1                  1
#> CBA4IHG       CBA4IHG stromal cell          1                  1
#> X129              129 stromal cell         22  0.363636363636364
#> X189              189 stromal cell         12  0.333333333333333
#> IFC               IFC stromal cell          3  0.666666666666667
#> X1234569      1234569 stromal cell          4               0.25
#> X7                  7 stromal cell          5                0.6
#> GFC               GFC stromal cell          5                0.6
#> X149              149 stromal cell          6  0.333333333333333
#> X3                  3 stromal cell         23  0.478260869565217
#> X5                  5 stromal cell          3  0.333333333333333
#> X789              789 stromal cell          5                0.4
#> X16I              16I stromal cell          3  0.333333333333333
#> X14789          14789 stromal cell          2                0.5
#> X12345          12345 stromal cell          3  0.333333333333333
#> C81               C81 stromal cell          4                0.5
#> X569              569 stromal cell          3  0.333333333333333
#> E                   E stromal cell          2                0.5
#> X749              749 stromal cell          2                0.5
#> X12349          12349 stromal cell          9  0.444444444444444
#> A23               A23 stromal cell          3  0.333333333333333
#> CB9               CB9 stromal cell          3  0.333333333333333
#> X3D9              3D9 stromal cell          2                0.5
#> X12389          12389 stromal cell          3  0.333333333333333
#>                        pvalue                fdr
#> I          0.0145160768151399 0.0145160768151399
#> X1         0.0608235969466868 0.0608235969466868
#> X9         0.0995901067413939 0.0995901067413939
#> X123        0.105983669882209  0.105983669882209
#> X1234789    0.214555812830038  0.214555812830038
#> X56IHG      0.214555812830038  0.214555812830038
#> C           0.379469229242591  0.379469229242591
#> X56789      0.379469229242591  0.379469229242591
#> X3FE        0.465317919075145  0.465317919075145
#> IFE         0.465317919075145  0.465317919075145
#> X12G        0.465317919075145  0.465317919075145
#> X145        0.465317919075145  0.465317919075145
#> X1256789    0.465317919075145  0.465317919075145
#> X12347      0.465317919075145  0.465317919075145
#> X123456789  0.465317919075145  0.465317919075145
#> X12C69      0.465317919075145  0.465317919075145
#> X12GFEDC89  0.465317919075145  0.465317919075145
#> CB14GFE     0.465317919075145  0.465317919075145
#> EDA         0.465317919075145  0.465317919075145
#> X7H9        0.465317919075145  0.465317919075145
#> CB56789     0.465317919075145  0.465317919075145
#> CBAHG       0.465317919075145  0.465317919075145
#> X1HEDCB9    0.465317919075145  0.465317919075145
#> X12345F9    0.465317919075145  0.465317919075145
#> C456IHG     0.465317919075145  0.465317919075145
#> X16G        0.465317919075145  0.465317919075145
#> X125        0.465317919075145  0.465317919075145
#> X123FED9    0.465317919075145  0.465317919075145
#> X12589      0.465317919075145  0.465317919075145
#> X3456789    0.465317919075145  0.465317919075145
#> X589        0.465317919075145  0.465317919075145
#> X12789      0.465317919075145  0.465317919075145
#> X14589      0.465317919075145  0.465317919075145
#> X12E        0.465317919075145  0.465317919075145
#> AD5         0.465317919075145  0.465317919075145
#> X567H9      0.465317919075145  0.465317919075145
#> X5HGBA      0.465317919075145  0.465317919075145
#> X1234589    0.465317919075145  0.465317919075145
#> X16I8G      0.465317919075145  0.465317919075145
#> X14IHGFE    0.465317919075145  0.465317919075145
#> A2349       0.465317919075145  0.465317919075145
#> CBA4IHG     0.465317919075145  0.465317919075145
#> X129        0.513439912481979  0.513439912481979
#> X189        0.558124870810537  0.558124870810537
#> IFC         0.597915933089743  0.597915933089743
#> X1234569    0.627362703495069  0.627362703495069
#> X7          0.664626433145562  0.664626433145562
#> GFC         0.664626433145562  0.664626433145562
#> X149        0.691282799193364  0.691282799193364
#> X3          0.828687570264039  0.828687570264039
#> X5                          1                  1
#> X789                        1                  1
#> X16I                        1                  1
#> X14789                      1                  1
#> X12345                      1                  1
#> C81                         1                  1
#> X569                        1                  1
#> E                           1                  1
#> X749                        1                  1
#> X12349                      1                  1
#> A23                         1                  1
#> CB9                         1                  1
#> X3D9                        1                  1
#> X12389                      1                  1